refactor(metrics): move ComputeSessionMetricsFromTrace to cmd/ to fix dependency violation
sim/cluster imported sim/workload, violating the cmd/ → sim/cluster/ → sim/ rule.
The function and its 5 behavioral tests move to cmd/session_metrics_trace.go/.._test.go;
observe_cmd.go calls the unexported local function directly.
