+++ cmd/observe_cmd.go
@@ -16,7 +16,6 @@ import (
 	"time"
 
 	"github.com/inference-sim/inference-sim/sim"
-	"github.com/inference-sim/inference-sim/sim/cluster"
 	"github.com/inference-sim/inference-sim/sim/workload"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -503,7 +502,7 @@ func runObserve(cmd *cobra.Command, _ []string) {
 	logrus.Infof("Trace exported: %d records to %s / %s", len(records), observeTraceHeader, observeTraceData)
 
 	// Print session metrics if any record carries a session label (#1058)
-	sessionMetrics := cluster.ComputeSessionMetricsFromTrace(records)
+	sessionMetrics := computeSessionMetricsFromTrace(records)
 	printSessionMetrics(os.Stdout, sessionMetrics)
 
 	// Export ITL if requested (BC-5: opt-in)
